+++ verl/workers/engine/megatron/transformer_impl.py
@@ -617,7 +617,7 @@ class MegatronEngineWithLMHead(MegatronEngine):
         multi_modal_inputs = model_inputs["multi_modal_inputs"]
 
         if not isinstance(temperature, torch.Tensor):
-            temperature = torch.tensor([temperature] * input_ids.shape[0], device=input_ids.device)
+            temperature = torch.tensor([temperature] * input_ids.shape[0], device=input_ids.device, dtype=torch.float32)
 
         assert temperature.shape[0] == input_ids.shape[0]
         temperature = verl_F.expand_as_nested(temperature, input_ids)  # (bsz, j1)
